[MOV] website_hr_recruitment: Split models files according to guidelines
