+++ addons/website_hr_recruitment/models/__init__.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+from . import hr_applicant
 from . import hr_department
-from . import hr_recruitment
+from . import hr_job
+from . import hr_recruitment_source
 from . import website
