+++ addons/website_hr_recruitment/models/hr_applicant.py
@@ -1,0 +1,23 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from odoo import models, _
+
+
+class Applicant(models.Model):
+
+    _inherit = 'hr.applicant'
+
+    def website_form_input_filter(self, request, values):
+        if 'partner_name' in values:
+            applicant_job = self.env['hr.job'].sudo().search([('id', '=', values['job_id'])]).name if 'job_id' in values else False
+            name = '%s - %s' % (values['partner_name'], applicant_job) if applicant_job else _("%s's Application", values['partner_name'])
+            values.setdefault('name', name)
+        if values.get('job_id'):
+            stage = self.env['hr.recruitment.stage'].sudo().search([
+                ('fold', '=', False),
+                '|', ('job_ids', '=', False), ('job_ids', '=', values['job_id']),
+            ], order='sequence asc', limit=1)
+            if stage:
+                values['stage_id'] = stage.id
+        return values
